use informers instead of watchers for default posture
Watchers are simple to implement and lowest level of abstraction to monitor K8s events but it's not resilient. SharedInformer provides eventually consistent linkage to the objects we want to monitor.

This commit fixes the issue where if due to some connection error we stopped watching namespace changes thereby breaking the namespace based default posture updates.

Signed-off-by: daemon1024 <barun.acharya@accuknox.com>