enforcer: tame apparmor recursive behaviour
AppArmor `**` rule doesn't include the directory itself.
Taking an example of ps and proc. ps tries to access the `proc` directory itself before reading the underlying directories. When we save allow ps -> /proc/**, it didn't include the /proc/ directory itself so we update the directory rule to include both directory itself and the recursive pattern identifier.

Signed-off-by: daemon1024 <barun.acharya@accuknox.com>